uikit merge from EaseIM

+++ Sources/EaseChatUIKit/Classes/UI/Components/Conversation/ViewModel/ConversationViewModel.swift
@@ -8,6 +8,7 @@
 import Foundation
 import AudioToolbox
 
+public let disturb_change = "EaseUIKit_do_not_disturb_changed"
 
 /// Bind service and driver
 @objc open class ConversationViewModel: NSObject {
@@ -94,7 +95,7 @@ import AudioToolbox
     
     /// Load the session in the local database and create one if it does not exist
     /// - Parameters:
-    ///   - profile: ``EaseProfileProtocol``
+    ///   - profile: ``ChatUserProfileProtocol``
     ///   - text: Welcome message.
     /// - Returns: ``ConversationInfo`` object.
     @objc(loadIfNotExistCreateWithProfile:type:text:)
@@ -131,6 +132,7 @@ import AudioToolbox
     deinit {
         destroyed()
     }
+
 }
 
 //MARK: - ConversationListActionEventsDelegate
@@ -150,8 +152,14 @@ extension ConversationViewModel: ConversationListActionEventsDelegate {
         for id in ids {
             if let conversation = ChatClient.shared().chatManager?.getConversationWithConvId(id) {
                 if conversation.type == .chat {
+                    if let userCache = ChatUIKitContext.shared?.userCache?[id],!userCache.nickname.isEmpty {
+                        continue
+                    }
                     privateChats.append(id)
                 } else {
+                    if let groupCache = ChatUIKitContext.shared?.groupCache?[id],!groupCache.nickname.isEmpty {
+                        continue
+                    }
                     groupChats.append(id)
                 }
             }
@@ -289,15 +297,20 @@ extension ConversationViewModel: ConversationListActionEventsDelegate {
     }
     
     @objc open func delete(info: ConversationInfo) {
-        self.service?.deleteConversation(conversationId: info.id) { [weak self] error in
+
+        DialogManager.shared.showAlert(title: "Delete Conversation Alert".chat.localize, content: "Delete warning".chat.localize, showCancel: true, showConfirm: true) { [weak self] _ in
+            UIViewController.currentController?.dismiss(animated: true)
             guard let `self` = self else { return }
-            if error != nil {
-                consoleLogInfo("onConversationSwipe delete:\(error?.errorDescription ?? "")", type: .error)
-            } else {
-                if let infos = ChatClient.shared().chatManager?.getAllConversations(true) {
-                    self.driver?.refreshList(infos: self.mapper(objects: infos))
+            self.service?.deleteConversation(conversationId: info.id) { [weak self] error in
+                guard let `self` = self else { return }
+                if error != nil {
+                    consoleLogInfo("onConversationSwipe delete:\(error?.errorDescription ?? "")", type: .error)
+                } else {
+                    if let infos = ChatClient.shared().chatManager?.getAllConversations(true) {
+                        self.driver?.refreshList(infos: self.mapper(objects: infos))
+                    }
+                    self.updateUnreadCount()
                 }
-                self.updateUnreadCount()
             }
         }
     }
@@ -431,7 +444,7 @@ extension ConversationViewModel: ConversationServiceListener {
             self.driver?.refreshList(infos: items)
             
             if infos.count < 11 || self.firstLoadConversation {
-                let requestCount = infos.count < 11 ? (infos.count - 1):10
+                let requestCount = items.count < 11 ? items.count:10
                 if requestCount > 0 {
                     self.requestDisplayProfiles(ids: items.prefix(upTo: requestCount).map({ $0.id }))
                     self.firstLoadConversation = false
@@ -446,7 +459,15 @@ extension ConversationViewModel: ConversationServiceListener {
     
     @objc open func conversationMessageAlreadyReadOnOtherDevice(info: ConversationInfo) {
         info.unreadCount = 0
-        self.service?.markAllMessagesAsRead(conversationId: info.id)
+        if let infos = ChatClient.shared().chatManager?.getAllConversations(true) {
+            let items = self.mapper(objects: infos)
+            var count = UInt(0)
+            for item in items where item.doNotDisturb == false {
+                count += item.unreadCount
+            }
+            self.service?.notifyUnreadCount(count: count)
+        }
+//        self.service?.markAllMessagesAsRead(conversationId: info.id)
         self.driver?.swipeMenuOperation(info: info, type: .read)
     }
 }
